HH.View.Utils/Converters/DoubleToPercentageConverter.cs
using System;
using System.Globalization;
using System.Windows.Data;

namespace HH.View.Utils.Converters
{
    public sealed class DoubleToPercentageConverter : IValueConverter
    {
        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
        {
            var doubleValue = (double)value;
            return doubleValue * 100;
        }

        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
        {
            throw new NotImplementedException();
        }
    }
}
